Visual Assist Update.

+++ code/qcommon/cvar.cpp
@@ -1291,7 +1291,7 @@ void	Cvar_Update(vmCvar_t* vmCvar) {
 	if (strlen(cv->string) + 1 > MAX_CVAR_VALUE_STRING)
 		Com_Error(ERR_DROP, "Cvar_Update: src %s length %u exceeds MAX_CVAR_VALUE_STRING",
 			cv->string,
-			(unsigned int)strlen(cv->string));
+			strlen(cv->string));
 	Q_strncpyz(vmCvar->string, cv->string, MAX_CVAR_VALUE_STRING);
 	vmCvar->value = cv->value;
 	vmCvar->integer = cv->integer;
