+++ code/rd-vanilla/tr_shade_calc.cpp
@@ -228,7 +228,7 @@ void RB_CalcBulgeVertexes(deformStage_t* ds)
 
 		for (i = 0; i < tess.numVertexes; i++, xyz += 4, st += 2 * NUM_TEX_COORDS, normal += 4)
 		{
-			const int off = (float)(FUNCTABLE_SIZE / (M_PI * 2)) * (st[0] * ds->bulgeWidth + now);
+			const int off = FUNCTABLE_SIZE / (M_PI * 2) * (st[0] * ds->bulgeWidth + now);
 
 			const float scale = tr.sinTable[off & FUNCTABLE_MASK] * ds->bulgeHeight;
