+++ codemp/botlib/be_ai_char.cpp
@@ -523,7 +523,7 @@ int BotInterpolateCharacters(int handle1, int handle2, float desiredskill)
 	strcpy(out->filename, ch1->filename);
 	botcharacters[handle] = out;
 
-	const float scale = (float)(desiredskill - ch1->skill) / (ch2->skill - ch1->skill);
+	const float scale = (desiredskill - ch1->skill) / (ch2->skill - ch1->skill);
 	for (int i = 0; i < MAX_CHARACTERISTICS; i++)
 	{
 		//
