+++ codemp/client/cl_keys.cpp
@@ -449,7 +449,7 @@ void Field_VariableSizeDraw(field_t* edit, int x, int y, int width, int size, qb
 
 	// draw the cursor
 	if (showCursor) {
-		if ((int)(cls.realtime >> 8) & 1) {
+		if (cls.realtime >> 8 & 1) {
 			return;		// off blink
 		}
