+++ codemp/qcommon/cvar.cpp
@@ -1506,7 +1506,7 @@ void	Cvar_Update(vmCvar_t* vmCvar) {
 	}
 	vmCvar->modificationCount = cv->modificationCount;
 	if (strlen(cv->string) + 1 > MAX_CVAR_VALUE_STRING)
-		Com_Error(ERR_DROP, "Cvar_Update: src %s length %u exceeds MAX_CVAR_VALUE_STRING", cv->string, (unsigned int)strlen(cv->string));
+		Com_Error(ERR_DROP, "Cvar_Update: src %s length %u exceeds MAX_CVAR_VALUE_STRING", cv->string, strlen(cv->string));
 	Q_strncpyz(vmCvar->string, cv->string, MAX_CVAR_VALUE_STRING);
 
 	vmCvar->value = cv->value;
