+++ codemp/qcommon/md5.cpp
@@ -166,7 +166,7 @@ void MD5Update(MD5Context* ctx, unsigned char const* buf, unsigned len)
 	/* Update bitcount */
 
 	uint32_t t = ctx->bits[0];
-	if ((ctx->bits[0] = t + ((uint32_t)len << 3)) < t)
+	if ((ctx->bits[0] = t + (len << 3)) < t)
 		ctx->bits[1]++;		/* Carry from low to high */
 	ctx->bits[1] += len >> 29;
