+++ codemp/rd-vanilla/tr_surface.cpp
@@ -742,7 +742,7 @@ static void RB_SurfaceCylinder(void)
 	// this doesn't need to be perfect....just a rough compensation for zoom level is enough
 	length *= backEnd.viewParms.fovX / 90.0f;
 
-	float detail = 1 - (float)length / 1024;
+	float detail = 1 - length / 1024;
 	int segments = NUM_CYLINDER_SEGMENTS * detail;
 
 	// 3 is the absolute minimum, but the pop between 3-8 is too noticeable
